导航不瞎转了，调自瞄，现在的想法是算出自瞄发的imu和云台imu的差值，然后与电机的当前角度相加或者做差，得到目标角度，然后计算pid，也就是继续后续云台的control

+++ PYRo/Module/Yaw/pyro_yaw.h
@@ -18,7 +18,7 @@ namespace pyro
 struct yaw_cmd_t : cmd_base_t
 {
     float target_yaw_imu_angle;
-    double radar_imu{};
+    float radar_imu{};
 
     // float test_yaw_radps;
